requirements.txt
pydantic>=2.8.0
playwright>=1.52.0
python-dotenv>=1.0.1
pytest>=8.0.0
requests>=2.32.0
